fix(server): move _pick_fallback_model to constants to avoid flask import in tests
The test_server_cli_fallback.py ran in the 'no extras' CI matrix which
doesn't have flask installed. Importing from gptme.server.cli triggered
gptme/server/__init__.py → app.py → flask, causing an ImportError.

Move _pick_fallback_model to gptme/server/constants.py (no flask dep),
and update the test import accordingly.
